svg/interact/inheritance.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="8000"><title>Inheritance of pointer-events</title><metadata><h:link href="https://svgwg.org/svg2-draft/interact.html#PointerEventsProperty" rel="help"/><h:meta content="pointer-events inherits according to the spec." name="assert"/><h:meta content="pointer-events has initial value auto." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/inheritance-testcommon.js"/><script><![CDATA[

assert_inherited('pointer-events', 'auto', 'none');

  ]]></script></svg>
